Add files via upload
In analysis module program is asking for a path to text to by analyzed. This part can be changed to a proper file uploading method. Similar situation is in analysis, when user is asked to input name of a file - it should be replaced with a method: choose from list.